+++ lib/src/countly_flutter.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io' show Platform;
 
+import 'package:countly_flutter/src/content_builder.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:pedantic/pedantic.dart';
@@ -211,6 +212,17 @@ class Countly {
           log('[FMethodCallH] $e', logLevel: LogLevel.ERROR);
         }
         break;
+      case 'contentCallback':
+        Map<String, dynamic> argumentsMap = Map<String, dynamic>.from(call.arguments);
+        final int contentResult = argumentsMap['contentResult'];
+        ContentStatus contentStatus = ContentStatus.completed;
+        if (contentResult == 1) {
+          contentStatus = ContentStatus.closed;
+        } 
+        Map<String, dynamic> contentData = Map<String, dynamic>.from(argumentsMap['contentData']);;
+
+        Countly.instance._contentBuilderInternal.onContentCallback(contentStatus, contentData);
+        break;
     }
   }
 
@@ -294,6 +306,11 @@ class Countly {
     }
     for (final callback in config.remoteConfigGlobalCallbacks) {
       Countly.instance._remoteConfigInternal.registerDownloadCallback(callback);
+    }
+
+    if(config.content.contentCallback != null){
+      log('[initWithConfig] About to register content callback', logLevel: LogLevel.VERBOSE);
+      Countly.instance._contentBuilderInternal.registerContentCallback(config.content.contentCallback!);
     }
 
     return result;